@@ -29,11 +29,11 @@ module.exports = {
     assetsPublicPath: '/',
     proxyTable: {
       '/login': {
-        target: process.env.AUTH_API_ADDRESS || 'http://127.0.0.1:8081',
+        target: process.env.AUTH_API_ADDRESS || 'http://127.0.0.1:8082',
         secure: false
       },
       '/todos': {
-        target: process.env.TODOS_API_ADDRESS || 'http://127.0.0.1:8082',
+        target: process.env.TODOS_API_ADDRESS || 'http://127.0.0.1:8083',
         secure: false
       },
       '/zipkin': {
