@@ -1,7 +1,9 @@
 import Vue from 'vue'
-
-import Auth from '@/auth'
 import Router from 'vue-router'
+import Auth from '@/auth'
+import Login from '@/components/Login.vue'
+import Todos from '@/components/Todos.vue'
+
 Vue.use(Router)
 
 export default new Router({
@@ -9,13 +11,13 @@ export default new Router({
     {
       path: '/login',
       name: 'login',
-      component: require('@/components/Login.vue')
+      component: Login
     },
     {
       path: '/',
       alias: '/todos',
       name: 'todos',
-      component: require('@/components/Todos.vue'),
+      component: Todos,
       beforeEnter: requireLoggedIn
     }
   ]
