@@ -31,15 +31,34 @@ public class UsersController {
     public User getUser(HttpServletRequest request, @PathVariable("username") String username) {
 
         Object requestAttribute = request.getAttribute("claims");
-        if((requestAttribute == null) || !(requestAttribute instanceof Claims)){
+        
+        // Allow access without JWT for inter-service communication
+        if (requestAttribute == null) {
+            System.out.println("DEBUG: No JWT claims found, allowing inter-service access for user: " + username);
+            return userRepository.findOneByUsername(username);
+        }
+
+        if (!(requestAttribute instanceof Claims)) {
             throw new RuntimeException("Did not receive required data from JWT token");
         }
 
         Claims claims = (Claims) requestAttribute;
 
-        if (!username.equalsIgnoreCase((String)claims.get("username"))) {
-            throw new AccessDeniedException("No access for requested entity");
+        String tokenUsername = (String) claims.get("username");
+        System.out.println("DEBUG: Token username = '" + tokenUsername + "', Requested username = '" + username + "'");
+        System.out.println("DEBUG: All claims = " + claims.toString());
+        
+        if (tokenUsername == null) {
+            System.out.println("DEBUG: Token username is null!");
+            throw new AccessDeniedException("Token does not contain username claim");
         }
+        
+        if (!username.equalsIgnoreCase(tokenUsername)) {
+            System.out.println("DEBUG: Username mismatch! Token: '" + tokenUsername + "', Requested: '" + username + "'");
+            throw new AccessDeniedException("No access for requested entity - Token username: " + tokenUsername + ", Requested: " + username);
+        }
+        
+        System.out.println("DEBUG: Authorization successful for user: " + username);
 
         return userRepository.findOneByUsername(username);
     }
