@@ -21,8 +21,12 @@ class HttpSecurityConfiguration {
 
         @Override
         protected void configure(HttpSecurity http) throws Exception {
-            http.antMatcher("/**")
-                    .addFilterAfter(jwtAuthenticationFilter, BasicAuthenticationFilter.class);
+            http.csrf().disable()
+                .authorizeRequests()
+                .antMatchers("/actuator/**", "/h2/**", "/health", "/users/**").permitAll()
+                .anyRequest().authenticated()
+                .and()
+                .addFilterAfter(jwtAuthenticationFilter, BasicAuthenticationFilter.class);
         }
     }
 }
